Labels: add right-alignment
Which changes the "pos" argument to specify the top-right instead of the top-left corner.

We then use this new property (together with max_width) to right-align the drink count on the profile page.

Doing this makes me slightly less angry at the IE developers for messing up the box model.