Introduce keycloak, but without actually using it
This adds keycloak to the development setup, and comes with Python code and sample usage of it on how it can be incorporated into the project. It doesn't yet replace LDAP, as we first need to move some LDAP dependencies to our database (i.e. timestamps stored in ldap about when the last time was that a user received a mail; we don't want this to have in keycloak, so instead we first need to refactor this to go to the local db).

Otherwise, the code is in principle working and allows us to fetch a list of users and their mail addresses. Credentials for the drinks-touch user are part of the example realm; no credentials have yet been generated for production.