ci(pre-commit): remove checkout limitation (#219)
Removes the checkout limitation for now, so that renovate PRs go green
again. We can discuss if this optimization is necessary here as next
step.

I think the default is to do a sparse checkout (only one commit) either
way. Is there even a reason to checkout more commits? :eyes:

Check the [commits of this
PR](https://github.com/flipdot/drinks-touch/pull/219/commits) for an
easier review, as I applied automatic formatting for all files. I made
manual changes, which I'll mark in a review below for you to verify.

---------

Co-authored-by: Soeren Wegener <wegener92@gmail.com>