Rendering: Implement caching & cache invalidation (#293)
Do not render ALL the things ALL the time. Instead, cache the surfaces
of elements and mark objects as dirty if properties that influence the
visual appearance change.

If you implement a `BaseElm` yourself, you may need to override
`BaseElm.calculate_hash` to include properties that influence the look
of your element.