Introduce "Unit of work" pattern
Instead of directly creating session objects everywhere, the unit of work pattern encapsulates multiple database operations into a single logical unit.

I've tried it out on the home screen, which is currently leaving idle transactions open.

The idea is that a function can be marked with a decorator as a "function which requires database access", and the decorator makes sure that a session object is passed to the function and the session is commited after the function has been called.

What do you think about this pattern? Does it make sense to apply it to the whole codebase?

My research: https://g.co/gemini/share/3262cf6849e1